mobile compat, file renaming, bugfixes
