--- data/events.js
@@ -1,4 +1,0 @@
-export default {
-    calendar_url: 'https://calendar.google.com/calendar/embed?src=acmbruins%40gmail.com&ctz=America%2FLos_Angeles&showPrint=0&showTitle=0&mode=AGENDA',
-    youtube_url: 'https://www.youtube.com/channel/UCwTqwgA4thysNbB2kNYsCIw',
-};
